js: move nodejs testing code to a separate package
